NDRS-324: minor update to get-balance endpoint
